# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SentinelX – Autonomous AI Red Team vs Blue Team Cyber Range API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /blue-team/alerts/{id}/acknowledge:
    patch:
      operationId: acknowledgeAlert
      tags: [blue-team]
      summary: Acknowledge a security alert
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alert acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecurityAlert"
  /blue-team/respond:
    post:
      operationId: executeBlueTeamStep
      tags: [blue-team]
      summary: Trigger the next autonomous Blue Team response step
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentStepInput"
      responses:
        "200":
          description: Response action result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentStepResult"
...
  /assets/{id}:
    get:
      operationId: getAsset
      tags: [assets]
      summary: Get a single network asset with details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Asset detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NetworkAsset"
        "404":
          description: Not found
  /assets/{id}/compromise:
    patch:
      operationId: updateAssetCompromise
      tags: [assets]
      summary: Update compromise status of an asset
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetCompromiseUpdate"
      responses:
        "200":
          description: Updated asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NetworkAsset"
...
  /incidents:
    get:
      operationId: listIncidents
      tags: [incidents]
      summary: List all security incidents
      responses:
        "200":
          description: Incident list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Incident"
    post:
      operationId: createIncident
      tags: [incidents]
      summary: Create a new incident case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentInput"
      responses:
        "201":
          description: Created incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
  /incidents/{id}:
    get:
      operationId: getIncident
      tags: [incidents]
      summary: Get incident details with timeline
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Incident detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "404":
          description: Not found
    patch:
      operationId: updateIncident
      tags: [incidents]
      summary: Update incident status or severity
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentUpdate"
      responses:
        "200":
          description: Updated incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
  /incidents/{id}/timeline:
    get:
      operationId: getIncidentTimeline
      tags: [incidents]
      summary: Get reconstructed incident timeline
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Timeline events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelineEvent"
  /incidents/{id}/playbook:
    post:
      operationId: runIncidentPlaybook
      tags: [incidents]
      summary: Run automated response playbook for an incident
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaybookInput"
      responses:
        "200":
          description: Playbook result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybookResult"

  # ── SOC Dashboard ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SimulationEvent:
      type: object
      required: [id, type, actor, description, severity, timestamp]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [attack, detection, response, compromise, containment, escalation]
        actor:
          type: string
          enum: [red_team, blue_team, commander, system]
        description:
          type: string
        severity:
          type: string
          enum: [info, low, medium, high, critical]
        assetId:
          type: ["integer", "null"]
        assetName:
          type: ["string", "null"]
        technique:
          type: ["string", "null"]
        mitreId:
          type: ["string", "null"]
        timestamp:
          type: string
...
    AgentStepInput:
      type: object
      properties:
        context:
          type: ["string", "null"]
        forcePhase:
          type: ["string", "null"]
    AgentStepResult:
      type: object
      required: [agent, action, reasoning, outcome, nextStep]
      properties:
        agent:
          type: string
        action:
          type: string
        reasoning:
          type: string
        outcome:
          type: string
          enum: [success, partial, failed, detected, blocked]
        nextStep:
          type: string
        technique:
          type: ["string", "null"]
        mitreId:
          type: ["string", "null"]
        affectedAssets:
          type: array
          items:
            type: string
        generatedEvents:
          type: array
          items:
            $ref: "#/components/schemas/SimulationEvent"

    # ── Red Team ────────────────────────────────────────────────────────────────
...
    SecurityAlert:
      type: object
      required: [id, title, severity, status, source, description, timestamp]
      properties:
        id:
          type: integer
        title:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [open, acknowledged, investigating, resolved, false_positive]
        source:
          type: string
          enum: [ids, siem, edr, firewall, honeypot, anomaly_detection]
        description:
          type: string
        affectedAsset:
          type: ["string", "null"]
        mitreId:
          type: ["string", "null"]
        confidence:
          type: number
        timestamp:
          type: string
        acknowledgedAt:
          type: ["string", "null"]
...
    NetworkAsset:
      type: object
      required: [id, name, type, zone, ipAddress, os, services, compromiseLevel, status]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [web_server, app_server, database, domain_controller, workstation, email_server, firewall, router]
        zone:
          type: string
          enum: [dmz, internal, restricted, management]
        department:
          type: ["string", "null"]
          enum: [hr, finance, it, executive, public, null]
        ipAddress:
          type: string
        os:
          type: string
        services:
          type: array
          items:
            type: string
        vulnerabilities:
          type: array
          items:
            type: string
        compromiseLevel:
          type: integer
          minimum: 0
          maximum: 100
        status:
          type: string
          enum: [healthy, targeted, compromised, isolated, offline]
        x:
          type: ["number", "null"]
        y:
          type: ["number", "null"]
    AssetCompromiseUpdate:
      type: object
      required: [compromiseLevel, status]
      properties:
        compromiseLevel:
          type: integer
          minimum: 0
          maximum: 100
        status:
          type: string
          enum: [healthy, targeted, compromised, isolated, offline]
...
    Incident:
      type: object
      required: [id, title, severity, status, description, affectedAssets, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [open, investigating, contained, resolved, closed]
        description:
          type: string
        affectedAssets:
          type: array
          items:
            type: string
        assignee:
          type: ["string", "null"]
        rootCause:
          type: ["string", "null"]
        mitigations:
          type: array
          items:
            type: string
        evidenceCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        resolvedAt:
          type: ["string", "null"]
    IncidentInput:
      type: object
      required: [title, severity, description]
      properties:
        title:
          type: string
          minLength: 1
        severity:
          type: string
          enum: [low, medium, high, critical]
        description:
          type: string
        affectedAssets:
          type: array
          items:
            type: string
    IncidentUpdate:
      type: object
      properties:
        status:
          type: ["string", "null"]
          enum: [open, investigating, contained, resolved, closed, null]
        severity:
          type: ["string", "null"]
          enum: [low, medium, high, critical, null]
        assignee:
          type: ["string", "null"]
        rootCause:
          type: ["string", "null"]
    TimelineEvent:
      type: object
      required: [id, timestamp, type, description, actor]
      properties:
        id:
          type: integer
        timestamp:
          type: string
        type:
          type: string
          enum: [attack, detection, response, escalation, containment, resolution]
        description:
          type: string
        actor:
          type: string
        evidence:
          type: ["string", "null"]
        assetId:
          type: ["integer", "null"]
    PlaybookInput:
      type: object
      required: [playbookType]
      properties:
        playbookType:
          type: string
          enum: [isolate, forensics, patch, reset_credentials, block_ip, notify_stakeholders]
    PlaybookResult:
      type: object
      required: [playbookType, steps, success, summary]
      properties:
        playbookType:
          type: string
        steps:
          type: array
          items:
            type: object
            required: [step, status, output]
            properties:
              step:
                type: string
              status:
                type: string
                enum: [completed, failed, skipped]
              output:
                type: string
        success:
          type: boolean
        summary:
          type: string

    # ── SOC ─────────────────────────────────────────────────────────────────────
...